Add files via upload
Bootstrapped maximum likelihood estimation for Poisson distribution using the gamma-arrivals dataset.  This is a proposed solution to the following problem: The file gamma-ray contains a small quantity of data collected from the Compton Gamma Ray Observatory, a satellite launched by NASA in 1991 (http://cossc.gsfc.nasa.gov/). For each of 100 sequential time intervals of variable lengths (given in seconds), the number of gamma rays originating in a particular area of the sky was recorded. Assuming a model that the arrival times are a Poisson process with constant emission rate (Î» = events per second), estimate Î». What is the estimated standard error? How might you informally check the assumption that the emission rate is constant? What is the posterior distribution of Î» if an improper gamma prior is used?